fix: move PermissionNotificationType
