+++ frontend/src/types/NotificationTypes.ts
@@ -1,3 +1,6 @@
+import type { MessageProps } from "primevue/message";
+import type { VNode } from "vue";
+
 export type SeverityType = "success" | "warn" | "error";
 
 export const Severity = {
@@ -10,4 +13,11 @@ export type ForestClientNotificationType = {
     type: "Duplicate" | "Error" | "NotExist" | "NotActive" | "Invalid";
     severity: SeverityType;
     clientNumbers: string[];
+};
+
+export type PermissionNotificationType = {
+    serverity: MessageProps["severity"];
+    message: string | VNode | (() => VNode);
+    hasFullMsg: boolean;
+    fullMessage?: string | VNode | (() => VNode);
 };
